requirements.txt
backports.ssl-match-hostname==3.5.0.1
docker-py==1.9.0
requests==2.11.1
six==1.10.0
websocket-client==0.37.0
